Code refactoring.
changed module name: render -> renderers
changed trait name: Rend -> Render
changed enum name: Render -> Renderer
