+++ src/lib.rs
@@ -1,8 +1,8 @@
 
 pub use pixels::Error;
-use render::{Render, Rend};
-use render::s2d::S2D;
-use render::s3d::S3D;
+use renderers::{Renderer, Render};
+use renderers::s2d::S2D;
+use renderers::s3d::S3D;
 use pixels::{Pixels, SurfaceTexture};
 use pixels::Error::Surface;
 use winit::dpi::LogicalSize;
@@ -12,7 +12,7 @@ use winit::window::{Window, WindowBuilder};
 use winit::window::CursorIcon;
 use winit_input_helper::WinitInputHelper;
 
-mod render;
+mod renderers;
 
 
 pub enum RenderType {
@@ -74,8 +74,8 @@ impl CanvasBuilder {
             height: h,
             title: self.canvas.title,
             render: match self.canvas.render {
-                RenderType::S2D => Render::S2D(S2D::new(w, h)),
-                RenderType::S3D => Render::S3D(S3D::new(w, h)),
+                RenderType::S2D => Renderer::S2D(S2D::new(w, h)),
+                RenderType::S3D => Renderer::S3D(S3D::new(w, h)),
             }
         }
     }
@@ -85,7 +85,7 @@ pub struct Canvas {
     width: u32,
     height: u32,
     title: String,
-    render: Render,
+    render: Renderer,
 }
 
 impl Canvas {
@@ -115,7 +115,7 @@ impl Canvas {
                 self.render.render(pixels.get_frame());
                 if pixels
                     .render()
-                    .map_err(|e| panic!("pixels.render() failed: {e}"))
+                    .map_err(|e| panic!("pixels.renderers() failed: {e}"))
                     .is_err()
                 {
                     *control_flow = ControlFlow::Exit;
