+++ src/renderers/s2d.rs
@@ -1,12 +1,17 @@
-use crate::render::Rend;
-use pixels::wgpu::Color;
 
-pub struct S3D {
+use crate::renderers::Render;
+use pixels::wgpu::Color;
+pub use crate::renderers::s2d::shape::Shape;
+
+mod shape;
+
+pub struct S2D {
+    shapes: Vec<Shape>,
     width: u32,
     height: u32,
 }
 
-impl Rend for S3D {
+impl Render for S2D {
     fn render(&self, screen: &mut [u8]) {
         for (i, pixel) in screen.chunks_exact_mut(4).enumerate() {
             let x = (i % self.width as usize) as i16;
@@ -17,34 +22,12 @@ impl Rend for S3D {
     }
 }
 
-impl S3D {
+impl S2D {
     pub fn new(width: u32, height: u32) -> Self {
         Self {
+            shapes: vec![],
             width: width,
             height: height,
         }
     }
-}
-
-struct Point {
-    x: f64,
-    y: f64,
-    z: f64
-}
-
-pub struct Camera {
-    eye: Point,
-    center: Point,
-    up: (i32, i32, i32)
-}
-
-pub enum Shader {
-    Flat,
-    Gouraud,
-    Phong
-}
-
-pub enum RenderMethod {
-    Rasterization,
-    RayTracing(i32)
 }
