+++ src/renderers/s2d/shape.rs
@@ -1,0 +1,5 @@
+
+pub enum Shape {
+    Shape(),
+
+}
