+++ src/renderers/s3d.rs
@@ -1,18 +1,12 @@
-
-use crate::render::Rend;
+use crate::renderers::Render;
 use pixels::wgpu::Color;
 
-enum Shape {
-
-}
-
-pub struct S2D {
-    shapes: Vec<Shape>,
+pub struct S3D {
     width: u32,
     height: u32,
 }
 
-impl Rend for S2D {
+impl Render for S3D {
     fn render(&self, screen: &mut [u8]) {
         for (i, pixel) in screen.chunks_exact_mut(4).enumerate() {
             let x = (i % self.width as usize) as i16;
@@ -23,12 +17,34 @@ impl Rend for S2D {
     }
 }
 
-impl S2D {
+impl S3D {
     pub fn new(width: u32, height: u32) -> Self {
         Self {
-            shapes: vec![],
             width: width,
             height: height,
         }
     }
+}
+
+struct Point {
+    x: f64,
+    y: f64,
+    z: f64
+}
+
+pub struct Camera {
+    eye: Point,
+    center: Point,
+    up: (i32, i32, i32)
+}
+
+pub enum Shader {
+    Flat,
+    Gouraud,
+    Phong
+}
+
+pub enum RenderMethod {
+    Rasterization,
+    RayTracing(i32)
 }
